+++ CHANGELOG.md
@@ -2,4 +2,4 @@
 
 ## v0.0.1
 
-- First implementation
+- Start template